Move file track functionality to separate package
